+++ src/main/java/com/bogus/carrental/model/dtos/CarUpdateDto.java
@@ -3,12 +3,14 @@ package com.bogus.carrental.model.dtos;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class CarUpdateDto {
 
     private Long mileage;
     private String color;
-    private Double payForDay;
+    private BigDecimal payForDay;
 
 }
